requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
onnxruntime==1.19.0
pillow==10.4.0
python-multipart==0.0.9
pydantic==2.9.0
pytest==8.3.0
httpx==0.27.0
aiofiles==24.1.0